server/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
psycopg2-binary
supabase
pyjwt
bcrypt
email-validator
python-multipart
requests